updated the post courses route to require basic auth when a course is created.
When a person trys to create a course, their un and pw is checked via basic auth.  If they're in the database the course they create will carry their user id.  This is a working solution but the code needs to be refactored using a more appropriate promise based approach.